Refactored all db code to separate file.
