+++ data_collection.py
@@ -29,4 +29,37 @@ for index, row in df.iterrows():
 #print(ad_ids)
 
 
-
+youtube_category_codes = {
+    1: 'Film',
+    2: 'Autos and Vehicles',
+    10: 'Music',
+    15: 'Pets and Animals',
+    17: 'Sports',
+    18: 'Short Movies',
+    19: 'Travel and Events',
+    20: 'Gaming',
+    21: 'Videoblogging',
+    22: 'People and Blogs',
+    23: 'Comedy',
+    24: 'Entertainment',
+    25: 'News and Politics',
+    26: 'HowTo and Style',
+    27: 'Education',
+    28: 'Science and Technology',
+    29: 'Nonprofits and Activism',
+    30: 'Movies',
+    31: 'Anime/Animation',
+    32: 'Action/Adventure',
+    33: 'Classics',
+    34: 'Comedy',
+    35: 'Documentary',
+    36: 'Drama',
+    37: 'Family',
+    38: 'Foreign',
+    39: 'Horror',
+    40: 'Sci-Fi/Fantasy',
+    41: 'Thriller',
+    42: 'Shorts',
+    43: 'Shows',
+    44: 'Trailers'
+}
